add task, user, users, refactor user and users to singleUser and users, hook into components
