+++ client/store/index.js
@@ -2,8 +2,8 @@ import {createStore, combineReducers, applyMiddleware} from 'redux'
 import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import user from './user'
-import task from './task'
+import singleUser from './singleUser'
+import tasks from './tasks'
 import singleTask from './singleTask'
 import usersReducer from './users'
 
@@ -18,4 +18,4 @@ const middleware = composeWithDevTools(
 const store = createStore(reducer, middleware)
 
 export default store
-export * from './user'
+export * from './singleUser'
