+++ client/store/users.js
@@ -1,43 +1,44 @@
 import axios from 'axios'
 
-/**
- * ACTION TYPES
- */
-
+const GET_ALL_USERS = 'GET_ALL_USERS'
 const UPDATE_USER = 'UPDATE_USER'
 
-/**
- * INITIAL STATE
- */
-const defaultUser = []
+const getAllUsers = (users) => ({
+  type: GET_ALL_USERS,
+  users,
+})
+const updateUser = (user) => ({
+  type: UPDATE_USER,
+  user,
+})
 
-/**
- * ACTION CREATORS
- */
-
-export const updateUser = (user) => {
-  return {
-    type: UPDATE_USER,
-    user,
+export const fetchAllUsers = () => {
+  return async (dispatch) => {
+    try {
+      const {data} = await axios.get('/api/users')
+      dispatch(getAllUsers(data))
+    } catch (err) {
+      console.error(err)
+    }
   }
 }
-
-/**
- * THUNK CREATORS
- */
-
 export const fetchUpdateUser = (user) => {
   return async (dispatch) => {
-    const {data} = await axios.put(`/api/users/${user.id}`, user)
-    dispatch(updateUser(data))
+    try {
+      const {data} = await axios.put(`/api/users/${user.id}`, user)
+      dispatch(updateUser(data))
+    } catch (err) {
+      console.error(err)
+    }
   }
 }
 
-/**
- * REDUCER
- */
-export default function usersReducer(state = defaultUser, action) {
+const initState = []
+
+export default (state = initState, action) => {
   switch (action.type) {
+    case GET_ALL_USERS:
+      return action.users
     case UPDATE_USER:
       return state.map((user) => {
         return user.id === action.user.id ? action.user : user
